+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Added
 - Support for `compartment_id` filter in `email_senders` and `email_suppressions` data sources
+- Support for import in dbHomes and dbSystems
 
 ### Fixed
 - Backward compatibility for compositeId in Object Storage - Objects and PARs
@@ -15,7 +16,6 @@
 - Support for provider service keys in Fast Connect Provider Services
 - Singular data sources for User, Group, File Storage Snapshot, Private IP and Virtual Cloud Network (VCN).
 - Support for authentication policy introduced in v3.18.0 is now generally available.
-- Support for import in dbHomes and dbSystems
 
 ### Fixed
 - Virtual Circuit update failures by handling default values